v2/cmd/benchstack: label top N phases overall
When we seriously shrink a phase, it currently just disappears from
the labels since it's so small in the right-most cell. Since it would
be nice to know what we shrank so much, this changes from using the
top N phases in the right-most cell to the top N overall.

One complication is that some of the top N overall phases may be not
be present in the right-most cell at all. Hence, this also constructs
a global phase order over all cells and uses that to order the labels.

Change-Id: I96aacad3f35c1e5930d1445182adc16579bf2b45