v2/cmd/benchstack: use graph coloring to assign phase colors
Currently we just assign phase colors in order of first observation,
but a lot of phases are really small and disappear visually, so we
wind up with identically-colored phases visually adjacent to each
other.

Switch to using a graph coloring that both separates visually large
and small phases and finds a coloring that works across all cells.

Change-Id: Ie8149d2d38b2c576ea17e021d3857c42b8166d19