v2/cmd/benchstack: delta visualization for live-B and heap-B
This adds a new type of visualization for heap sizes. Since these can
have both positive and negative deltas, the existing stack
visualization isn't sufficient.

This visualization makes it easy to see the peak, as well as which
phases contribute the most to the peak. It's also fairly easy to
compare between cells, though this isn't the focus of this
visualization.

Change-Id: Ie1019b7611a8e416eb067ca4c3004b5c5ca87c7f