v2/benchproc: complete redesign to use schemas
This replaces the schema-less approach with a schema-based approach.
Now, a projection describes a schema, how to map from benchfmt.Results
to configs with that schema, and how to sort configs produced by that
schema. While this is somewhat less flexible than the current
schema-less system (for example, there's no way to directly construct
a Config in the new API), it leads to a dramatically simpler API.
Furthermore, since the new Configs are logically flat, it also leads
to dramatically simpler implementations of sorting and prefix merging.

Change-Id: Ibb2a6993cfea85891a0a86c7a1c57cd8b0d2e3b8